TIP-860: Move ProductModel related api tests
